chore(dev) auto-convert Windows line-endings
Git clients on Windows tend to convert line endings to Windows
format, such that Windows editors can be used (is reverted on commit).
This doesn't work if mounting those files from a Unix machine for
development for example (Kong-Vagrant). Lua is agnostic to the format,
but the Unix shell is not. Hence only the scripts in ./kong/bin are
affected.

This change will have the makefile automatically convert those scripts
back to Unix format. This allows Windows users to work on the Kong code
and custom plugins without other configuration.